Fase Final

+++ trunk/TFC_final/web/ConsultaMunicipios.jsp
@@ -12,7 +12,7 @@
         <script src="js/functions.js" type="text/javascript"></script>
 
         <!--Titulo-->
-        <title>Cadastro de Ocorrências</title>
+        <title>Consulta de Municípios</title>
         <%@taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%> 
         <%@taglib uri="http://java.sun.com/jsp/jstl/sql" prefix="sql"%> 
 
@@ -49,136 +49,71 @@
     </head>
     <body>
         <div id="container">
-            <header>
-                <div id="header">
-                    <div id="logo"><a href="Principal.jsp"><img src="images/logo.png" alt="peterson" /></a></div>
-                    <div id="wrapper-menu">
-                        <ul id="nav" name="nav">
-                            <li id="item1" class="fade"><a class="cadastro-basico"  href="#" 
-                                                           title="Cadastros BÃ¡sicos"><img src="images/woofunction-icons/folder_add_32.png" width="20" height="20" 							                alt="cadastro" /> Cadastro</a></li>
+            <div id="header">
+                <div id="logo"><a href="Principal.jsp"><img src="images/logo.png" alt="peterson" /></a></div>
+                <div id="wrapper-menu">
+                    <ul id="nav" name="nav">
+                        <li id="item1" class="fade"><a class="cadastro-basico"  href="#"
+                                                       title="Cadastros Básicos"><img src="images/woofunction-icons/folder_add_32.png" width="20" height="20" alt="cadastro" /> Cadastro</a></li>
+                        <li id="item2" class="fade"><a  class="cadastro-basico2" href="#" title="Consultas Gerenciais"><img src="images/woofunction-icons/folder_chart_32.png" width="20" height="20" alt="consulta" /> Consultas</a></li>
+                        <li id="item3" class="fade"><a href="relatorio.jsp" title="Impressão de Relatórios"><img src="images/woofunction-icons/folder_page_32.png" width="20" height="20" alt="relatorio" /> Relatórios</a></li>
+                        <li id="item5" class="fade"><a href="ajuda.jsp"     title="Portal de Ajuda"><img src="images/woofunction-icons/folder_warning_32.png" width="20" height="20" alt="ajuda" />Ajuda</a></li>
+                    </ul>
+                    <!--end nav--> 
 
-                            <li id="item2" class="fade"><a  class="cadastro-basico2" href="#" title="Consultas Gerenciais"><img src="images/woofunction-icons/folder_chart_32.png" width="20" height="20" alt="consulta" /> Consultas</a></li>
+                    <!--Inicio - Menu de Cadastros-->
+                    <div id="big-menu-hidden">
+                        <div id="product-list"> <span class="btn-close"><a class="cadastro-basico" href="#">close</a></span>
+                            <h2 id="li-item1" title="SEO KEYWORDS"><strong>Cadastros Básicos</strong></h2>
+                            <ul id="cadastro">
+                                <li><a href="CadastroBairro.jsp" title="Cadastro de novos Bairros">Bairros</a></li>
+                                <li><a href="CadastroMunicipio.jsp" title="Cadastro de Municípios">Municípios</a></li>
+                                <li><a href="CadastroAnimais.jsp" title="Cadastro de Animais">Animais</a></li>
+                                <li><a href="CadastroOcorrencias.jsp" title="Cadastro de Ocorrência">Ocorrência</a></li>
+                                <li class="view-all"></li>
+                            </ul>
 
-                            <li id="item3" class="fade"><a class="cadastro-basico3" href="#" title="ImpressÃ£o de RelatÃ³rios"><img src="images/woofunction-icons/folder_page_32.png" width="20" height="20" alt="relatorio" /> Relatorios</a></li>
+                            <ul id="Ocorrencia">
+                                <li><a href="CadastroSituacao.jsp" title="Cadastro de Situação de Ocorrências">Situação</a></li>
+                                <li><a href="CadastroTipo.jsp" title="Cadastro de Tipo de Ocorrências">Tipo</a></li>
+                                <li><a href="CadastroMateriais.jsp" title="Cadastro de Materiais">Materiais</a></li>
 
-                            <li id="item5" class="fade"><a href="ajuda.jsp" title="Portal de Ajuda"><img src="images/woofunction-icons/folder_warning_32.png" width="20" height="20" alt="ajuda" />Ajuda</a></li>
-                        </ul><!--end nav-->
-
-                        <!--Inicio - Menu de Cadastros-->
-                        <div id="big-menu-hidden">
-                            <div id="cart">
-                                <div id="counter-cart">
-                                    <div id="btn-go-cart">	                
-                                        <h3 id="li-item1" >Cadastro de OcorrÃªncias</h3>
-                                    </div>
-                                    <p id="btn-go-cart"><a href="" title="Cadastro de novos Bairros">Bairro</a></p>
-                                    <p id="btn-go-cart"><a href="" title="Cadastro de materiais apreendidos por OcorrÃªncia">ApreensÃ£o</a></p>
-                                    <p id="btn-go-cart"><a href="" title="Cadastro de novos MunicÃ­pios">Municipio</a></p>
-                                </div>
-                            </div>
-                            <div id="product-list">
-                                <h2 id="li-item1" >Outros <strong>Cadastros:</strong></h2>
-                                <span class="btn-close"><a class="cadastro-basico"href="#">close</a></span>
-                                <ul id="list-cameras">
-                                    <li><a href="endereco.jsp" title="Cadastro de Localidade">Localidade</a></li>
-                                    <li><a href="materiais.jsp" title="Cadastro de Materiais">Materiais</a></li>
-                                    <li><a href="os.jsp" title="Cadastro de Constatação de Ordem de Serviço">ConstataÃ§Ã£o</a></li>
-                                    <li><a href="#" title="Cadastro de Infraï¿½ï¿½o">Auto de InfraÃ§Ã£o</a></li>
-                                    <li><a href="#" title="Cadastro de Recolhimento">Recolhimento</a></li>
-                                    <li><a href="os.jsp" title="Cadastro de Ordem de ServiÃ§o">Ordem de ServiÃ§o</a></li>
-                                    <li class="view-all"><span><a href="#"></a></span></li>
-                                </ul>
-                            </div>
-                        </div><!--Fim Menu de Cadastros-->
-
-                        <!--Inicio Menu de RelatÃ³rios-->
-
-                        <div id="big-menu-hidden3">
-                            <div id="cart">
-                                <div id="counter-cart">
-                                    <div id="btn-go-cart">
-                                        <h3 id="li-item1" >Tipo de OcorrÃªncia</h3>
-                                    </div>
-                                    <ul id="list-cameras">
-                                        <li><a href="" title="">Fauna99</a></li>
-                                        <li><a href="" title="">Flora99</a></li>
-                                    </ul>
-                                </div>
-                            </div>
-
-                            <div id="product-list">
-                                <div id="cart">
-                                    <div id="counter-cart">
-                                        <div id="btn-go-cart">
-                                            <h3 id="li-item1" >Atendimentos</h3>
-                                            <ul id="list-cameras">
-                                                <li><a href="" title="teste">Pendentes</a></li>
-                                                <li><a href="" title="teste">Atendidos</a></li>
-                                                <li class="view-all"><span><a href="#"></a></span></li>
-
-                                            </ul>
-                                        </div>
-                                    </div>
-                                </div>
-
-                            </div>
-                        </div><!--Fim Menu de relatÃ³rios-->
-                        <!--Inicio menu de Consultas-->
-                        <div id="big-menu-hidden2">
-                            <div id="cart">
-                                <div id="counter-cart">
-                                    <div id="btn-go-cart">	                
-                                        <h3 id="li-item1" > Consultas</h3>
-                                    </div>
-                                    <p id="btn-go-cart"><a href="" title="222Cadastro referente a Flora">222Flora</a></p>
-                                    <p id="btn-go-cart"><a href="" title="222Cadastro referente a Fauna">222Fauna</a></p>
-                                    <p id="btn-go-cart"><a href="" title="222Cadastro referente a Ordem Publica">222Ordem</a></p>
-
-                                </div>
-                            </div>
-                            <div id="product-list">
-
-                                <h2 id="li-item1" >Outros <strong>Cadastros:</strong></h2>
-                                <span class="btn-close"><a class="cadastro-basico2"href="#">close</a></span>
-                                <ul id="list-cameras">
-
-                                    <li><a href="endereco.jsp" title="aaaaCadastro de Localidade">222Localidade</a></li>
-                                    <li><a href="materiais.jsp" title="aaaaaCadastro de Materiais">22Materiais</a></li>
-                                    <li><a href="os.jsp" title="aaaaCadastro de Constataï¿½ï¿½oo de Ordem de Serviï¿½o">222ConstataÃ§Ã£o</a></li>
-                                    <li><a href="#" title="aaaaaCadastro de Infraï¿½ï¿½o">222Auto de Infraï¿½ï¿½o</a></li>
-                                    <li><a href="#" title="2222Cadastro de Recolhimento">222Recolhimento</a></li>
-                                    <li><a href="os.jsp" title="Cadastro de Ordem de ServiÃ§o">222Ordem de Serviï¿½ï¿½o</a></li>
-                                    <li class="view-all"><span><a href="#"></a></span></li>
-                                </ul>
-
-                            </div>
-                        </div><!--Fim Menu de Consultas-->
-
-                        <!-- Inicio Menu de Materiais-->
-                        <div id="big-menu-hidden4">
-                            <div id="cart">
-                                <div id="counter-cart">
-                                    <div id="btn-go-cart">	                
-                                        <h3 id="li-item1" > Consultas</h3>
-                                    </div>
-
-                                </div>
-                            </div>
-                            <div id="product-list">
-
-                                <h2 id="li-item1" >Outros <strong>Cadastros:</strong></h2>
-                                <span class="btn-close"><a class="cadastro-basico4"href="#">close</a></span>
-                                <ul id="list-cameras">
-
-
-                                </ul>
-
-                            </div>
+                                <li class="view-all"></li>
+                            </ul>
                         </div>
-                        <!-- Fim Menu de Materiais-->                        
-
                     </div>
+                    <!--Fim Menu de Cadastros--> 
+                    <!--Inicio menu de Consultas-->
+                    <div id="big-menu-hidden2"> 
+                        <!--  <div id="cart">
+                          <div id="counter-cart">
+                          <p>0 items in</p>
+                          <p id="btn-go-cart"><a href="" title="keywords for SEO">YOUR CART</a></p>
+                         <span id="img-empty-cart"><img src="images/ico-empty-cart.png" alt="keyword for SEO" width="109" height="131"/></span> </div>
+                      </div>-->
+                        <div id="product-list"> <span class="btn-close"><a class="cadastro-basico2" href="#">close</a></span>
+                            <h2 id="li-item1" title="SEO KEYWORDS"><strong>Consultas</strong></h2>
+                            <ul id="Consultas">
+                                <li><a href="ConsultaBairros.jsp" title="Consulta de Bairros">Bairros</a></li>
+                                <li><a href="ConsultaMunicipios.jsp" title="Consulta de Municípios">Municípios</a></li>
+                                <li><a href="ConsultaOcorrencias.jsp" title="Consulta de Ocorrências">Ocorrências</a></li>
+                                <li><a href="ConsultaTipo.jsp" title="Consulta de Tipos da Ocorrência">Tipo</a></li>
+                                <li class="view-all"></li>
+                            </ul>
+
+                            <ul id="Consultas2">
+                                <li><a href=" .jsp" title="Consulta de Materiais Apreendidos">Materiais</a></li>
+                                <li><a href="consultaAnimais.jsp" title="Consulta de Animais Apreendidos">Animais</a></li>
+                                <li><a href="ConsultaSituacao.jsp" title="Consulta de Situações da Ocorrência">Situação</a></li>
+                                <li><a href="ConsultaOS.jsp" title="Consulta de Ordens de Serviço">Ordem de Serviço</a></li>
+                                <li class="view-all"></li>
+                            </ul>
+                        </div>
+                    </div>
+                    <!--Fim Menu de Consultas-->  
+
                 </div>
+            </div>
             </header>
             <div id="content">
                 <div id="content_head"></div>
@@ -213,7 +148,7 @@
                                             <c:forEach var="municipio" items="${municipio.rows}">
                                                 <tr>
                                                     <td class="td1"><a href="CadastroMunicipio.jsp?id=${municipio.COD_MUN}&acao=edit"><img src="images/woofunction-icons/pencil_32.png" value="submit" width="20" height="20"/></a></td>
-                                                    <td class="td1"><a href="ConsultaMunicipios.jsp?id=${municipio.COD_MUN}&acao=delete"><img src="images/woofunction-icons/close_16.png" value="submit" width="20" height="20"/></a></td>
+                                                    <td class="td1"><a href="ConsultaMunicipios.jsp?id=${municipio.COD_MUN}&acao=delete" onclick="return confirm('Deseja excluir o registro?')"><img src="images/woofunction-icons/close_16.png" value="submit" width="20" height="20"/></a></td>
                                                     <td class="td1"><c:out value="${municipio.COD_MUN}"/></td>
                                                     <td class="td1"><c:out value="${municipio.DES_MUN}"/></td>
